LU-4209 utils: fix O_LOV_DELAY_CREATE conflict
In kernel 3.11 O_TMPFILE was introduced, but the open flag value
conflicts with the O_LOV_DELAY_CREATE flag 020000000 added to fix
LU-812 in Lustre 2.4.  Instead of trying to find a non-conflicting
flag for O_LOV_DELAY_CREATE, use (O_NOCTTY|FASYNC) as the new flag
value.  These flag are not meaningful for newly-created Lustre files
and should be fine since O_LOV_DELAY_CREATE is only meaningful for
regular files and directories.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I565f3454616edc60c6acee01034aa5d773500c1e