LU-4249 quota: fix race in setting lqe_lockh
When the per-ID lock is being canceled, the lqe_lockh will be cleared,
however, a just received reply of pending DQACQ can set the lqe_lockh
back to stale lockh, that'll cause the pre-acquire fails for -ENOLCK
constantly.

This patch fix the qsd_req_completion() to only set the lqe_lockh
on enqueueing intent per-ID lock but not on the normal DQACQ request.

Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Change-Id: Id60e1a4106ca733a6b4845e8ed1fa70775c8242d