LU-4298 utils: do not create file with no striping info
'lfs setstripe' should not create a file with no striping
information if bad values are given.

We are now checking that the command-line input value of
the offset index is within the number of OSTs in the file
system.

Caching is added to llapi_search_fsname to speed up calls
to lfs, like 'lfs setstripe', with multiple files on the
command line.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Signed-off-by: Sebastien Buisson <sebastien.buisson@bull.net>
Change-Id: I19beef973dc3a62bf2cb443e02195896772702fc