LU-3582 test: Add run_test to Runtests
There is no run_test in this script so add one
so logging in Autotest will greatly improve.
It seems some of the test is dependant so I did a
very minimal change to get the better logging.

The patch is back-ported from the following patch:
Lustre-commit: c74be05495cd5b7489fdbb100cd7b6d15db32053
Lustre-change: http://review.whamcloud.com/7014

Signed-off-by: Keith Mannthey <keith.mannthey@intel.com>
Change-Id: I96d7fe464d81094842dc5939519d74f7168d96b1
Signed-off-by: Jian Yu <jian.yu@intel.com>