LU-3326 tests: sync after file deleted in run_*.sh
In run_*.sh, call wait_delete_completed() after the file created
is deleted, then the dirty pages of the deleted file will be
flushed to the OST if MDT has not sent the object destroy request
to it.

This patch also fixes the issue of missing LDLM_FL_DISCRAD_DATA
when sending blocking AST of LDLM extent lock.

Signed-off-by: Hongchao Zhang <hongchao.zhang@intel.com>
Signed-off-by: Jian Yu <jian.yu@intel.com>
Change-Id: Ia246122a5c3471b7cbea9cdaa7fad2ea4db856c5