LU-4015 ldiskfs: access external journal using device path
Currently ldiskfs when configured with an external journal
uses the device number to access it. The problem with this
approach is that the device number can change across
reboots. This patch adds a new mount option -o journal_path
to tell ldiskfs where the external journal is instead. This
patch is based on commit landed to what will be Lustre 2.6
which is d406aa3e71cde588ea53d42ff2596d77e65e33f8.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I86af198cdab54164583ae1c03fff891edfc1f152