LU-4287 autoconf: Improve test for proc write and read functions
For newer kernels the proc api has changed to using only
seq_files. To detect this I tested the presence of PDE_DATA
to determine if it was a newer kernel. Now in RHEL6.5 this
macro has been backported but the rest of the changes have
not so it has broken the build. This patch moves from
testing PDE_DATA to testing if you can call write_proc_t
in struct proc_dir_entry. In newer kernels that function
is missing.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I96d21ae5f055d5077da69acdcc4948ec160dd63a