LU-4154 lfsck: skip old lfsck test in DNE mode
The old e2fsck/lfsck tool will not be allowed to run on a DNE
filesystem. This patch updates generate_db() to pass master MDS
parameters only, so that the old lfsck does not corrupt it or
delete all of the files on other MDTs.
This patch also fixes a typo in run_lfsck_remote().

Test-Parameters: mdscount=2 mdtcount=2 testlist=lfsck mdsfilesystemtype=ldiskfs mdtfilesystemtype=ldiskfs ostfilesystemtype=ldiskfs
Signed-off-by: Emoly Liu <emoly.liu@intel.com>
Change-Id: I28caaa5ca4a564aabbd6116325257c905ff22861