LU-2995 osd: Delete the ref of remote parent in del ..
Delete remote dir from REMOTE PARENT during delete .., instead
of final object destroying(osd_object_destroy), because the
object might not be destroyed if it is being opened by someone
else, instead it will be inserted to the ORPHAN dir, so if it
does not delete the {.., REMOTE_PARENT}, it will be two parents
for this remote dir.

Add remote object FID into op_data->op_fid3, so the stale check
will not fail during remote open.

Signed-off-by: wang di <di.wang@intel.com>
Change-Id: I6bbac9174869ae02f385ac52740bb2046f0f1bab