LU-4371 tests: remove hard-coded UID values
To make testing more robust in different test configs, remove the
hard-coding of UID 500/501 in sanity-gss.sh and get this from the
environment if available.  Don't try to grep it out of /etc/passwd,
in case it is stored in LDAP or some other auth database.

Consolidate the UID != RUNAS_ID check in check_runas_id().

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I76da9b7d682aad576f7be91587c802c678f1ee54