LU-4471 mdd: mdd_unlink: do trans_start after sanity check
Currently, mdd_trans_start is called before
mdd_unlink_sanity_check.  This means a remote directory
which has files in it can be removed on MDT0 before the
sanity check on MDT1 finds the files and errors, which
orphans those files on MDT1.

This patch moves mdd_trans_start after the sanity check.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I08882b682b9f0016577214821efec4759ee5c184