LU-3946 ptlrpc: Allow to force crash-dump vs log upon evict
This patch is intended to provide a Lustre-Client 2.4.0-RC2 distro
with the ability to force LBUG/crash-dump, instead of debug log flush,
upon eviction. If [obd_]dump_on_eviction = -1 LBUG/crash-dump
(depending on panic_on_lbug value).
This was required as part of LU-3946 debugging where there is a need
to get a crash-dump to be generated upon eviction in order to
understand the root cause of the problem, and particularly if it is
related to LU-874 and some kind of dead-lock in LDLM, or comes from
some tricky issue (scheduling/overload, like in LU-4112).

Test-Parameters: fortestonly testlist=recovery-small
Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>
Change-Id: I1f7ecbdeb9040c69eb0d584b8e6165ed10f892ef