TEI-000 test-infrastructure, Upgrade Test Build
This is a master upgrade test build. It can be used to test master
against test profiles such as el6.5.

Instructions for use.

rebase this patch.
edit the comment in git so the distributions below are the ones
you want to test.

You will probably need some manual, retest with jiralogger to
ensure it runs on all test clusters.

Wait for the results to appear in Maloo, they should all pass
because this is master.

Test-Requirements: clientdistro=el6.5 serverdistro=el6.5

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: Id2befd61531901e2080d4b54cf01082b61dcaa9f