LU-4582 mgc: replace hard-coded MGC_ENQUEUE_LIMIT value
During client mount, the client will send an LDLM_ENQUEUE request to
MGS with delay limit set to MGC_ENQUEUE_LIMIT, which is hard coded to
50 seconds. On the other hand, the interval for pinger is deduced from
obd_timeout. When obd_timeout is configured for a longer period of
time, so does the pinger. We know that connecting to the secondary MGS
node is triggered by the pinger. Now that we have a longer interval,
the pinger will not be able to try the secondary before the
LDLM_ENQUEUE request fails the mount using the same delay limit.

This code change will assoicate that specific LDLM_ENQUEUE req delay
to obd_timeout, instea of a hard-coded value, such that the client can
have a chance to connect to the secondary.

Xyratex-bug-id: MRP-1516
Signed-off-by: Cheng Shao <cheng_shao@xyratex.com>
Change-Id: I70a27574f7b1480f85d19840bc556551efb4dff7