LU-459 tests: quiet spurious console/test messages
Quiet a number of frequent and useless messages printed to the test
output and/or the console.  These distract from understanding what
is happening in the test and provide little, if any, value.

The client "acl" mount option has been deprecated since 1.8.0, as
it was added in mid-1.6, and is driven by the MDS config only.

Use textual debug/subsystem flags instead of numeric flags.  This
has been possible through all of 1.8, but is deprecated in 2.x so
we can move to a more sophisticated debug mechanism in the future.

Lustre-commit: 26bb579a96efced8825b63b84465334368ecc382
Lustre-change: http://review.whamcloud.com/2015
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I81bed13c22465dc179e19f5a6714b732d63ebbe5