LU-3606 fallocate: Implement fallocate() support for Lustre
This patch adds fallocate() support for Lustre. fallocate() method
of the inode_operations is implemented and transported to the OSTs
to interface with the underlying OSD's fallocate() code.

FALLOC_FL_PUNCH_HOLE functionality uses the OST_PUNCH RPC. A new RPC,
OST_PREALLOC has been added for space preallocation.

The fallocate functionality (both punch and prealloc) in CLIO has
been multiplexed with CIT_SETATTR. The punch code builds on top of
the truncate code in such a way that truncate is treated as a special
case of punch at the EOF.

Signed-off-by: Swapnil Pimpale <spimpale@ddn.com>
Change-Id: I03f27d356616fbf3a3ab8e6309af26c00434d81b