LU-4647 nodemap: add idmap functions to mdt
Add functions to allow the mdt to use the nodemap module
to map UIDs and GIDs. When a client connects to the mdt,
the NID of the client is classified by nodemap and a pointer
to the nodemap is added to the export structure.

In old_init_ucred and old_init_ucred_reint, the body ids and fs
ids are mapping according to the nodemap state (enabled, maps,
squashed ids, trust_client_ids and allow_root_access).

This does not implement mappings ACLs (extended attributes) yet.

Signed-off-by: Joshua Walgenbach <jjw@iu.edu>
Change-Id: I9648f6837f8bb4ff145c8a0d647fc934c08aaffe