LU-4656 tests: Add MOUNT_EXTRA for extra mount options
This patch adds a MOUNT_EXTRA variable which can be set to any extra
options to the mount command, such as "-n".  MOUNTOPT will be reserved
for a "-o opts" list of options, while MOUNT_EXTRA will be for
everything else.

MOUNT_EXTRA was added to the mount commands and output messages in
zconf_mount and zconf_mount_clients in test-framework.sh.  MOUNT_EXTRA
was given an empty default value in cfg/local.sh, and short
documentation was added above MOUNTOPT and MOUNT_EXTRA.

Two lines in test_132 were also edited to use "lctl get_param -n"
instead of piping the output of "lctl get_param" through an awk
command.

Signed-off-by: Ryan Haasken <haasken@cray.com>
Change-Id: I32d078fc6b64ee93199ad7b881791e872590f711