LU-4495 ldlm: distributed deadlock with clio.
Early sublock allocation enter to situation with practically cached
locks. In that case we may hit deadlock when client1 linked a sub1,
sub2 as new sublocks, but sub3 as cached sublock, so client2 will hold
sub1,sub2 and tries to enqueue a sub3.. so it's deadlock as releasing
a sub3 blocked by enqueue a sub1.

Xyratex-bug-id: CLSTR-2003
Signed-off-by: Patrick Farrell<paf@cray.com>
Signed-off-by: Alexey Lyashkov <alexey_lyashkov@xyratex.com>
Change-Id: I8c792081b415464451b014584e7188a47757f955