LU-4020 hsm: allow copytool event monitoring with JSON
Adds hooks into various llapi_hsm_* functions to emit JSON-formatted
event messages for consumption by a monitoring agent. The copytool
needs to be supplied with an optional --event-fifo argument to
enable this feature.

WIP: Pushing for proof of concept, but should be reasonably close
to production-quality, barring major concerns with design.

Also, put all JSON routines in a separate file/lib to allow for
LGPL licensing.

Signed-off-by: Michael MacDonald <michael.macdonald@intel.com>
Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>
Change-Id: I1fecfb8e60ef091c6d37d16dc1faf40be45c5ee2