LU-4647 nodemap: add mapping functionality
Add functions to allow the MDT to use the nodemap module
functions to map UIDs and GIDs. When a client connects to the MDT,
the NID of the client is classified by nodemap and a pointer
to the nodemap is added to the export structure.
In old_init_ucred and old_init_ucred_reint, the body ids and fs
ids are mapping according to the nodemap state (enabled, maps,
squashed ids, trust_client_ids and allow_root_access). ACLs are mapped
in mdt_getattr_internal. Squashed ACLs are discarded before the ACLs
are sent to clients. Only trusted administrative nodes are allowed to
setfacl.

Change-Id: I9648f6837f8bb4ff145c8a0d647fc934c08aaffe
Signed-off-by: Joshua Walgenbach <jjw@iu.edu>
Signed-off-by: Kit Westneat <kit.westneat@gmail.com>