LU-4879 lfsck: avoid check exp_connect_flags before connected
It is possible that the connection among some MDTs and OSTs are not
ready yet when the LFSCK wants to start the scanning globally. Under
such case, the LFSCK RPC should trigger related initial connection or
recovery connection, but not check the exp_connect_flags() directly.

On the other hand, it is not important to check whether the remote
server to support the OBD_CONNECT_LFSCK or not, because if the peer
does not support the LFSCK, the lfsck_start RPC will get failure.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I3b881dea23ccb7695d57ec422538f731fc77e657